% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{agent_set}
\alias{agent}
\alias{agent_del}
\alias{agent_get}
\alias{agent_has}
\alias{agent_set}
\title{Token Agent}
\usage{
agent_set(name, value)

agent_has(name)

agent_get(name)

agent_del(name)
}
\arguments{
\item{name}{a unique name to store and retrieve this object. Make sure to
pick a unique name that does not conflict with other packages using the same
store.}

\item{value}{an R object to be stored securely. This part contains the
sensitive data.}
}
\description{
Safely store and load sensitive data. This API may be both by users as well
as packages to store e.g. secrets or auth tokens.
}
\details{
Note that it is up to the user to secure the keyring with a password via
\link{update_password}.
}

