% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{new_agent}
\alias{new_agent}
\title{Build an agent object}
\usage{
new_agent(attributes)
}
\arguments{
\item{attributes}{A named list.}
}
\value{
An agent object.
}
\description{
Build an agent object
}
\examples{
new_agent()
new_agent(list(energy = 10, health = 20))
}
