% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password.R
\name{update_password}
\alias{reset_password}
\alias{update_password}
\title{Update Password}
\usage{
update_password()

reset_password()
}
\description{
Your keyring is only secure when protected by a good password. These functions
can only be called interatively by the user (not by packages).
}
\details{
Updating the password automatically refreshes all keypairs. If you forgot your
password you can reset it with `reset_password`. This will permanently delete
all your existing tokens.
}

