% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{agent}
\alias{agent}
\title{Build an agent object}
\usage{
agent(attributes = list())
}
\arguments{
\item{attributes}{A named list.}
}
\value{
An agent object.
}
\description{
Build an agent object
}
\examples{
agent()
agent(list(energy = 10, health = 20))
}
